+++ frontend/src/pages/Admin/EarnStamp/components/EarnStampSkeleton/index.tsx
@@ -1,0 +1,13 @@
+import { SkeletonContainer, SkeletonCoupon, SkeletonHeader, SkeletonSubHeader } from './style';
+
+const EarnStampSkeleton = () => {
+  return (
+    <SkeletonContainer>
+      <SkeletonHeader />
+      <SkeletonSubHeader />
+      <SkeletonCoupon />
+    </SkeletonContainer>
+  );
+};
+
+export default EarnStampSkeleton;
